LayoutTests/platform/ios-sim/http/tests/security/contentSecurityPolicy/media-src-blocked-expected.txt

CONSOLE MESSAGE: Refused to load the media '' because it violates the following Content Security Policy directive: "media-src 'none'".

CONSOLE MESSAGE: Refused to load the media '' because it violates the following Content Security Policy directive: "media-src 'none'".

CONSOLE MESSAGE: Refused to load the media 'http://127.0.0.1:8000/resources/load-and-stall.cgi?name=../../../media/content/test.mp4&mimeType=video/mp4&stallAt=100000' because it violates the following Content Security Policy directive: "media-src 'none'".

CONSOLE MESSAGE: Refused to load the media '' because it violates the following Content Security Policy directive: "media-src 'none'".

END OF TEST
This test passes if it doesn't alert failure.
